AshPolicyAuthorizer (ash_policy_authorizer v0.16.5)

An authorization extension for ash resources.

For more information, see AshPolicyAuthorizer.Authorizer

Types

request()

Specs

request() :: Ash.Engine.Request.t()

Functions

default_access_type(resource)

describe_resource(resource)

log_policy_breakdowns()

Whether or not ash policy authorizer is configured to show policy breakdowns in error messages

policies(resource)

show_policy_breakdowns?()

Whether or not ash policy authorizer is configured to show policy breakdowns in error messages

strict_check(actor, query, api)

A utility to determine if a given query/changeset would pass authorization.

This is still experimental.